Add detailed diagnose screen
Introduced a whole screen dedicated to show the details of a diagnose (including answers). Modifications on MyDiagnoses were necessary to add the press functionality. Added a couple more options to the generic components, including a title font and a option to ignore the base style of the button (as it is not necessary for images and some other options)