JB/gallery screen (#109)
* Gallery screen implementation

* Add permissions management

* Fix catch exception on not selected image

* Change font to raleway-bold, use theme colors in styles, move permissions functionality to permissionsService and fix imports.

* Refactor code and fix new photo flickering in gallery

Flickering was caused by a very specific problem regarding CellRenderer keys. Only solution was to use flexWrap property for contentContainerStyle in FlatList, which causes a REALLY annoying warning in __DEV__. I added the necessary logic to turn that warning off.

There still is a problem, in __PROD__ the screen works wonderfully, but in __DEV__ it's kind of slow. Further research will be done.

Co-authored-by: MPucheta <35865469+MPucheta@users.noreply.github.com>